requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
base58==2.1.0
    # via -r requirements.in
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.7
    # via requests
click==8.0.1
    # via -r requirements.in
idna==3.2
    # via requests
oauthlib==3.1.1
    # via requests-oauthlib
pycryptodome==3.10.1
    # via -r requirements.in
pyyaml==5.4.1
    # via -r requirements.in
requests==2.26.0
    # via
    #   -r requirements.in
    #   requests-oauthlib
    #   tweepy
requests-oauthlib==1.3.0
    # via tweepy
tweepy==4.1.0
    # via -r requirements.in
urllib3==1.26.7
    # via requests
